+++ include/Menus/ExtraSettingsMenu.hpp
@@ -27,11 +27,11 @@ public:
 protected:
 	void set_output_option(int index, int action);
 	bool is_option_selectable(int index, int action);
-	int get_num_options();
+	size_t get_num_options();
 	std::string get_string_option(int index, int action);
 	void class_setup();
 private:
 	int *options_indexes;
-	int num_enabled_options;
+	size_t num_enabled_options;
 };
 #endif
