+++ include/Menus/ScalingRatioMenu.hpp
@@ -31,12 +31,12 @@ public:
 protected:
 	bool is_option_inc_dec(int index);
 	void set_output_option(int index, int action);
-	int get_num_options();
+	size_t get_num_options();
 	std::string get_string_option(int index, int action);
 	void class_setup();
 private:
 	int *options_indexes;
-	int num_enabled_options;
+	size_t num_enabled_options;
 	std::string setTextOptionDualPercentage(int index, float value_1, float value_2);
 };
 #endif
