+++ source/Menus/ActionSelectionMenu.cpp
@@ -14,8 +14,8 @@ void ActionSelectionMenu::class_setup() {
 	this->width_divisor_menu = 9;
 	this->base_height_factor_menu = 12;
 	this->base_height_divisor_menu = 6;
-	this->min_text_size = 0.3;
-	this->max_width_slack = 1.1;
+	this->min_text_size = 0.3f;
+	this->max_width_slack = 1.1f;
 	this->menu_color = sf::Color(30, 30, 60, 192);
 	this->title = "Action Choice";
 	this->show_back_x = true;
@@ -33,7 +33,7 @@ void ActionSelectionMenu::reset_output_option() {
 }
 
 void ActionSelectionMenu::set_output_option(int index, int action) {
-	if(index >= this->get_num_options())
+	if(index >= ((int)this->get_num_options()))
 		this->selected_index = ACTION_SELECTION_MENU_BACK;
 	else if(index == BACK_X_OUTPUT_OPTION)
 		this->selected_index = ACTION_SELECTION_MENU_BACK;
@@ -41,7 +41,7 @@ void ActionSelectionMenu::set_output_option(int index, int action) {
 		this->selected_index = index;
 }
 
-int ActionSelectionMenu::get_num_options() {
+size_t ActionSelectionMenu::get_num_options() {
 	return this->possible_actions->size();
 }
