+++ source/Menus/AudioMenu.cpp
@@ -77,8 +77,8 @@ void AudioMenu::class_setup() {
 	this->width_divisor_menu = 9;
 	this->base_height_factor_menu = 12;
 	this->base_height_divisor_menu = 6;
-	this->min_text_size = 0.3;
-	this->max_width_slack = 1.1;
+	this->min_text_size = 0.3f;
+	this->max_width_slack = 1.1f;
 	this->menu_color = sf::Color(30, 30, 60, 192);
 	this->title = "Audio Settings";
 	this->show_back_x = true;
@@ -88,8 +88,8 @@ void AudioMenu::class_setup() {
 
 void AudioMenu::insert_data() {
 	this->num_enabled_options = 0;
-	for(int i = 0; i < NUM_TOTAL_MENU_OPTIONS; i++) {
-		this->options_indexes[this->num_enabled_options] = i;
+	for(size_t i = 0; i < NUM_TOTAL_MENU_OPTIONS; i++) {
+		this->options_indexes[this->num_enabled_options] = (int)i;
 		this->num_enabled_options++;
 	}
 	this->prepare_options();
@@ -108,7 +108,7 @@ void AudioMenu::set_output_option(int index, int action) {
 		this->selected_index = pollable_options[this->options_indexes[index]]->out_action;
 }
 
-int AudioMenu::get_num_options() {
+size_t AudioMenu::get_num_options() {
 	return this->num_enabled_options;
 }
 
@@ -142,7 +142,7 @@ void AudioMenu::prepare(float menu_scaling_factor, int view_size_x, int view_siz
 				this->labels[index]->setText(this->setTextOptionInt(real_index, audio_data->get_real_volume()));
 				break;
 			case AUDIO_MENU_MAX_LATENCY_DEC:
-				this->labels[index]->setText(this->setTextOptionInt(real_index, audio_data->get_max_audio_latency()));
+				this->labels[index]->setText(this->setTextOptionInt(real_index, (int)audio_data->get_max_audio_latency()));
 				break;
 			case AUDIO_MENU_OUTPUT_DEC:
 				this->labels[index]->setText(this->setTextOptionString(real_index, audio_data->get_audio_output_name()));
