+++ source/Menus/ColorCorrectionMenu.cpp
@@ -14,8 +14,8 @@ void ColorCorrectionMenu::class_setup() {
 	this->width_divisor_menu = 9;
 	this->base_height_factor_menu = 12;
 	this->base_height_divisor_menu = 6;
-	this->min_text_size = 0.3;
-	this->max_width_slack = 1.1;
+	this->min_text_size = 0.3f;
+	this->max_width_slack = 1.1f;
 	this->menu_color = sf::Color(30, 30, 60, 192);
 	this->title = this->base_name;
 	this->show_back_x = true;
@@ -43,7 +43,7 @@ void ColorCorrectionMenu::set_output_option(int index, int action) {
 		this->selected_index = index;
 }
 
-int ColorCorrectionMenu::get_num_options() {
+size_t ColorCorrectionMenu::get_num_options() {
 	return (*this->possible_correction_data).size();
 }
